Modify save map code
Now, the map is saved only if the it was modified (the new JSON is different to the old JSON, if any).

+++ core/src/main/kotlin/ro/luca1152/gravitybox/components/game/MapComponent.kt
@@ -148,7 +148,12 @@ class MapComponent : Component, Poolable {
         val fileFolder = "maps/editor"
         val existentFileName = getMapFileNameForId(levelId)
         if (existentFileName != "") {
-            Gdx.files.local("$fileFolder/$existentFileName").delete()
+            val oldJson = Gdx.files.local("$fileFolder/$existentFileName").readString()
+            if (oldJson == json.prettyPrint(json.writer.writer.toString())) {
+                return
+            } else {
+                Gdx.files.local("$fileFolder/$existentFileName").delete()
+            }
         }
         val fileHandle = Gdx.files.local("$fileFolder/${getNewFileName()}")
         fileHandle.writeString(json.prettyPrint(json.writer.writer.toString()), false)
